requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
future==1.0.0
idna==3.10
pymongo==4.10.1
python-dotenv==1.0.1
pyzk==0.9
requests==2.32.3
urllib3==2.2.3
pytz==2023.3
pandas==2.2.3
python-docx==1.1.2
openpyxl==3.1.5
hikvision-isapi-wrapper==0.3.4